Allow rule set to not write catalog file when only opened for reading (rule processing) Add trace information to default values Add tag information to missing facts Include meta data for default for functions 'missing' and 'exists' Fix function number to allow numbers to be passed to it instead of just strings. Fix g_month_day to ensure that month and day are always two digits.
[git-p4: depot-paths = "//ConsistencySuite/Server_Xule_Processor_DEV-Phillip/": change = 20567]

+++ xule/XuleRuleSet.py
@@ -76,6 +76,9 @@ class XuleRuleSet(object):
     
     def add(self, parseRes, file_time=None, file_name=None):
         
+        if self._openForAdd == False:
+            raise XuleRuleSetError(_("Attempting to add rule file, but rule set is not open for add"))
+        
         file_num = self._addXuleFile(file_time, file_name)
         
         # update the catalog
@@ -371,14 +374,14 @@ class XuleRuleSet(object):
     
         return pickle_name
                 
-    def open(self, ruleSetLocation):
+    def open(self, ruleSetLocation, open_for_add=True):
         #self.name = os.path.splitext(os.path.basename(ruleSetLocation))[0]
         self.location = ruleSetLocation
         try:
             with open(os.path.join(ruleSetLocation,"catalog.pik"),"rb") as i:
                 self.catalog = pickle.load(i)
             self.name = self.catalog['name']
-            self._openForAdd = True
+            self._openForAdd = open_for_add
         except FileNotFoundError:
             print("Cannot open catalog.")
             raise
@@ -480,24 +483,27 @@ class XuleRuleSet(object):
         return
         
     def addTaxonomy(self, taxonomy_location, entry_point):
-        base_location = os.path.join(self.location, 'base_taxonomy')
-        
-        if not os.path.isdir(taxonomy_location):
-            raise XuleRuleSetError(_("Taxonomy location '%s' is not a directory" % taxonomy_location))
-        
-        if not os.path.isfile(os.path.join(taxonomy_location, entry_point)):
-            raise XuleRuleSetError(_("Taxonomy entry point '%s' is not in '%s'" % (entry_point, taxonomy_location)))
-        
-        if os.path.exists(base_location):
-            if not os.path.isdir(base_location):
-                raise XuleRuleSetError(_("Taxonomy location '%s' is not a directory" % base_location))
-            else:
-                shutil.rmtree(base_location)
-
-        #copy the taxonomy location to the rule set
-        shutil.copytree(taxonomy_location, base_location)
-        
-        self.catalog['rules_dts_location'] = entry_point
+        if self._openForAdd == True:
+            base_location = os.path.join(self.location, 'base_taxonomy')
+            
+            if not os.path.isdir(taxonomy_location):
+                raise XuleRuleSetError(_("Taxonomy location '%s' is not a directory" % taxonomy_location))
+            
+            if not os.path.isfile(os.path.join(taxonomy_location, entry_point)):
+                raise XuleRuleSetError(_("Taxonomy entry point '%s' is not in '%s'" % (entry_point, taxonomy_location)))
+            
+            if os.path.exists(base_location):
+                if not os.path.isdir(base_location):
+                    raise XuleRuleSetError(_("Taxonomy location '%s' is not a directory" % base_location))
+                else:
+                    shutil.rmtree(base_location)
+    
+            #copy the taxonomy location to the rule set
+            shutil.copytree(taxonomy_location, base_location)
+            
+            self.catalog['rules_dts_location'] = entry_point
+        else:
+            raise XuleRuleSetError(_("Attempting to add taxonomy but the rule set is not open for add"))
             
     def getRulesTaxonomyLocation(self):
         if self.catalog['rules_dts_location'] is not None:
